[style | #39] 유저모달 스타일 수정
