+++ src/components/input/InputMed.vue
@@ -28,7 +28,7 @@ defineEmits(['update:modelValue'])
 </script>
 <style scoped>
 .input-med {
-  width: 18.75rem;
+  width: 17.5rem;
   height: 3.75rem;
   border-radius: 1rem;
   padding: 1.25rem 1rem;
